Show rendering canvases in the top docking area
Rendering canvases, as well as the data container inspector, are now
displayed as tabbed docks in the top docking area. The two docks that
were previously located in the right docking area have been moved to the
left to make the layout more intuitive.

We don't use the main window's central widget, but Qt requires it to
exist. This causes certain sizing problems, e.g. docks located in the
top and bottom docking areas don't expand to take up all the vertical
space, which results in some empty space in the centre of the window.
This will hopefully be fixed soon.

git-svn-id: https://camplinux.in.tum.de/svn/campvis/branches/gui-redux@506 bb408c1c-ae56-11e1-83d9-df6b3e0c105e