Fix issues with QtCanvas' sizing
The QtCanvas class didn't previously implement the sizeHint method. As
a result, Qt would allot it no display space if there were other widgets
competing for it (e.g. the main window's central widget).

Now QtCanvas has a sizeHint implementation that returns a size hint that
is in line with what the canvas' getSize method reports. Moreover,
QtCanvas now uses the `Expanding` size policy to inform Qt that it would
like to get as much display space as possible.

git-svn-id: https://camplinux.in.tum.de/svn/campvis/branches/gui-redux@519 bb408c1c-ae56-11e1-83d9-df6b3e0c105e