Support dragging VP widgets back into the MDI area
VisualizationPipelineWidget and PipelineMdiSubwindow have been extended
to report changes in their positions to VisualizationPipelineWrapper.
Based on the information the latter receives, it decides when to undock
MDI subwindows and dock floating pipeline widgets.

As a result, it's now possible to drag visualization pipeline widgets
back into the MDI area and turn them into MDI subwindows.