PipelineTreeWidget: provide a sensible height hint
The sizeHint method has been overridden in PipelineTreeWidget to return
a sensible height for the wigdet, while still reusing QTreeView's
hardcoded width.

The main window makes use of that to limit PipelineTreeWidget's maximum
height and give any vertical space reclaimed in this way over to the
widget displaying pipeline properties.