FloatPropertyWidget: use DoubleAdjusterWidget
Previously, FloatPropertyWidget used standard spin boxes and sliders to
display and let users modify the value of its associated property.
Consequently, it had to synchronise the states of those widgets on its
own.

Now that we have a separate widget that implements the two ways of
changing double values that FloatPropertyWidget provides, it makes sense
to simplify FloatPropertyWidget's code by making it use
DoubleAdjusterWidget internally.