Use DoubleAdjusterWidget in vector property widgets
Previously, widgets for double-valued vector properties used only spin
boxes to display and let users modify the values of their associated
property.

Now that we have DoubleAdjusterWidget, vector property widgets have been
updated to use it to make it possible to change the values of their
components with not only a spin box, but also a slider.