AbstractAdjusterWidget: base class for adjusters
AbstractAdjusterWidget is intended to be a base class for widgets
combining a slider with a spin box. It abstracts common functionality of
such widgets, leaving only signal/slot support and behaviour tweaks to
be added in concrete adjuster classes.

AbstractAdjusterWidget is a template class that can be parametrised with
either double or int. Its API closely follows that of QSpinBox/QSlider.