IntAdjusterWidget: widget for modifying ints
IntAdjusterWidget displays and provides two ways of modifying integers:
a slider that can be used to quickly change numeric values, and a spin
box, which is better suited for precise adjustments.

IntAdjusterWidget is a simple extension of AbstractAdjusterWidget<int>.
It is designed to be used by IntPropertyWidget and widgets for vector
properties storing integers.