Use IntAdjusterWidget in int and vector<int> property widgets
Previously, widgets for int and vector<int> properties used only spin
boxes to display and let users modify the values of their associated
property.

Now that we have IntAdjusterWidget, IntPropertyWidget and all
IVecPropertyWidgets have been updated to use it to make it possible to
change the values of their components with not only a spin box, but also
a slider.