LogHighlighter: fix compilation errors with GCC
Due to its language extensions, MSVC is more permissive than GCC with
respect to C++ code that doesn't fully comply with the language
standard. LogHighlighter relied on 2 such extensions: non-const rvalue
references, and closing nested template argument lists with '>>'
(instead of '> >'). Its code has now been fixed to conform to the
standard, and compiles fine with GCC 4.7 as a result.