Refactor VisualizationPipelineWrapper and rename it MdiDockArea
VisualizationPipelineWrapper has been renamed MdiDockArea, and refactored
to make it easier to use it with arbitrary widgets. It now inherits from
QMdiArea, which removes an unnecessary layer of indirection. Moreover, it
creates MDI subwindows and floating windows only if necessary, i.e. when
a widget stored in it changes state.