Fix bugs with the positioning of MDI windows
Previously, when an MDI window was being dragged very fast and its state
changed from docked to floating (or vice versa), it wouldn't be placed
under the cursor, but instead some way behind it (with regards to the
drag direction). This happened because we decided where to place the
window's based on its last position in its previous state. Now we use
the cursor's position to calculate the window's new position, which
fixes the problem.