Fix a double free in LogHighlighter's destructor
In its destructor, LogHighlighter first of all deleted its filter regex
explicitly, and then called `setFilterRegExp(0)`, triggering a second —
this time somewhat implicit — delete. If the filter regex pointer was
not NULL, this caused the application to crash due to a double-free.

Fix this error by not invoking `setFilterRegExp` in LogHighlighter's
destructor; the explicit delete performs the same cleanup as
`setFilterRegExp` anyways.