MdiDockableWindow: override the activateWindow method
The default implementation of activateWindow didn't forward activation
requests to MdiDockableWindow's sub-windows. Since MdiDockableWindow is
always hidden, the method was pretty much a no-op in disguise.

The overridden version of activateWindow properly forwards activation
requests, using the right method to activate the current sub-window, no
matter if it's floating or docked.