Dock DataContainerInspectorWidget in the MDI area
DataContainerInspectorWidget was previously stored in a regular dock
widget, but because of its rather large dimensions it didn't fit well in
any of the docking areas.

Put DataContainerInspectorWidget in an MdiDockableWindow and add it to
the MDI area. It fits much better there, and can still be undocked if
need be.