MDI: intercept only relevant move events
Both MdiDockedWindow and MdiFloatingWindow intercepted not only those
move events that were sent to them directly, but also the ones that the
widgets they hosted didn't handle. This made it possible to drag both
types of windows by clicking on an empty space in their widgets and
moving the mouse.

Ignore move events whose position is inside a widget hosted in a
MdiDockedWindow or MdiFloatingWindow. This makes interacting with such
widgets more predictable.