Enable the Vis module by default
No modules were previously enabled by default. As a result, the
application wasn't very useful in the default configuration.

Enable the Vis module by default to make the CAMPVis application
functional out of the box. Additional modules can be marked as default
in the future using the mechanism (SET_DEFAULT_MODULES) introduced in
this commit.

Closes #144