Initial support for pipelines implemented in Lua
A new class, LuaPipeline, has been added to the scripting module. It
provides a way of interfacing with pipelines implemented in Lua.

For the time being only the most basic operations (i.e. init and deinit)
are supported, but work on making it possible to implement
fully-functional pipelines in Lua is underway.

References #1