Lua bindings: wrap and instantiate tgt::Vector4
TGT-related glue code has been moved to a new interface file: tgt.i. It
wraps and instantiates tgt::Vector2 and (starting with this commit)
tgt::Vector4.

Many methods of those classes and a lot of typedefs haven't been wrapped
yet. This will be fixed when necessary in future commits.

References #1