Disable BUILD_SHARED_LIBS before processing liblua
When Lua is included in the build, its CMakeLists.txt enables
BUILD_SHARED_LIBS if it isn't explicitly disabled. This breaks the build
as CMake tries and fails to build all of CAMPVis' libraries as shared
libraries.

Disable BUILD_SHARED_LIBS before including `ext/lua` to prevent that.
Please note that this is a temporary fix; we'd like to eventually build
Lua and at least some of CAMPVis' libraries as shared libraries.

References #1