Use Lua helper classes to simplify LuaPipeline
LuaPipeline used to create a Lua VM itself and interact with it
directly. That caused its code to be unnecassarily bloated and complex as
it dealt with rather low-level details.

Delegate the management of Lua VM's state to the recently introduced
LuaVmState class and use the new table helpers to simplify LuaPipeline.

References #1