Finalise the implementation of the first Lua pipeline
Now that TGT and campvis-core can be built as shared libraries, the few
remaining issues with the first Lua pipeline could be tracked and fixed.
As a result, the pipeline can be executed without any issues, and its
functionality exactly corresponds to that of ResamplingDemo. To achieve
this, several additional classes, methods and constants had to be
wrapped.

References #1