Initial support for using Lua functions as slots
Certain types of pipelines need to react to changes in their constituent
processors. CAMPVis uses a signal-slot mechanism (based on the sigslot
library) for this purpose. Up until now, however, there was no way to
connect to a signal from a pipeline defined in Lua.

This commit adds initial support for connecting Lua functions to
sigslot's signals. The current implementation only works with unary
signals, and should be considered a proof of concept. To test it,
AbstractProcessor's s_validated signal has been wrapped using SWIG.

References #1