Implemented normalizing conversion between different GenericImageRepresentationLocal<BASETYPE, NUMCHANNELS> representations as proof-of-concept. This conversion might not always yield expected results since it remaps the entire input data type range to the entire output data type range. While this is reasonable for downcasts (narrowing conversions) a histogram stretch for widening conversions is probably not expected or wanted by the caller.
A more flexible and elaborate conversion as well as automatic conversion from ITK and OpenGL representations will need a redesign of the conversion interface since we're fighting with severe cyclic include dependency problems here. The future design should provide the possibility to externally register conversion functions for new data types that are not necessarily known in the core package (i.e. ITK representations).

refs #474