Better error message when Lua submodule is missing
If CAMPVIS_BUILD_LIB_LUA was set but the Lua Git submodule wasn't checked
out, the user would get a generic "could not find CMakeLists.txt" error
message which wasn't of much help.

Log a custom error message that provides a hint how to fix the problem in
the above situation.

References #1