Remove `scripting/testapp.cpp`
`scripting/testapp.cpp` contains a simple console application that was
used to test CAMPVis' scripting layer before it was ready for
integration with the main GUI application. Now that Lua pipelines can be
registered and executed the same way regular ones are, the test
application is no longer needed. Consequently, this commit removes it
from the project.

References #1